backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
alembic==1.14.0
psycopg2-binary==2.9.10
email-validator==2.1.1
azure-storage-blob==12.20.0

